refactor: move calc_action_log_probs to utils of model
