+++ applications/Chat/coati/experience_maker/naive.py
@@ -1,6 +1,6 @@
 import torch
 from coati.models.generation import generate_with_actor
-from coati.models.utils import compute_reward, normalize
+from coati.models.utils import calc_action_log_probs, compute_reward, normalize
 
 from .base import Experience, ExperienceMaker
 
@@ -24,9 +24,9 @@ class NaiveExperienceMaker(ExperienceMaker):
         num_actions = action_mask.size(1)
 
         actor_output = self.actor(sequences, attention_mask)
-        action_log_probs = self.actor.calc_action_log_probs(actor_output, sequences, num_actions)
+        action_log_probs = calc_action_log_probs(actor_output, sequences, num_actions)
         base_model_output = self.initial_model(sequences, attention_mask)
-        base_action_log_probs = self.initial_model.calc_action_log_probs(base_model_output, sequences, num_actions)
+        base_action_log_probs = calc_action_log_probs(base_model_output, sequences, num_actions)
         value = self.critic(sequences, action_mask, attention_mask)
         r = self.reward_model(sequences, attention_mask)
         reward = compute_reward(r, self.kl_coef, action_log_probs, base_action_log_probs, action_mask=action_mask)
